+++ src/budget/client/Overview/Modal/AdjustTxnModal.jsx
@@ -21,7 +21,7 @@ class AdjustTxnModal extends React.Component {
     const transaction = {
       transactionTypeGuid: this.refs.transactionTypeGuidSelect.value,
       accountGuid: this.refs.accountGuidSelect.value,
-      amount: parseInt(this.refs.amountInput.value),
+      amount: parseFloat(this.refs.amountInput.value),
       timestamp: moment(this.refs.timestampInput.value, ['MM|DD|YY']).format()
     }
