+++ src/budget/client/Overview/Modal/ResolvePlannedTxnModal.jsx
@@ -15,7 +15,7 @@ class ResolvePlannedTxnModal extends React.Component {
     const transaction = {
       transactionTypeGuid: this.refs.transactionTypeGuidSelect.value,
       accountGuid: this.refs.accountGuidSelect.value,
-      amount: parseInt(this.refs.amountInput.value),
+      amount: parseFloat(this.refs.amountInput.value),
       timestamp: moment(this.refs.timestampInput.value, ['MM|DD|YY']).format(),
       plannedTransactionGuid: this.props.plannedTxn.guid
     }
